cfu-service: dedupe common response helpers
